sentinelog-pro/backend-api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
elasticsearch[async]==8.16.0
pydantic==2.10.3
python-multipart==0.0.19
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis[asyncio]==5.0.1
